1. Code moved to root repository. (for compatibility with `bin/vendors`) 2. Change DI extension for loading xml configurations form Resources/config/*. (Example: twig.extension -> templating_twig.xml) 3. Changed the names of sources to resolve the conflict with source "array".
