--- DataGridBundle/Resources/views/blocks.html.twig
@@ -1,106 +1,0 @@
-{# ------------------------------------------------------ qrid ------------------------------------------------------ #}
-{% block grid %}
-<div class="grid">
-    <form action="{{ grid.routeUrl }}" method="post">
-        {% if grid.actions.count > 0 %}
-        {{ grid_massactions(grid) }}
-        {% endif %}
-        <table style="width:100%">
-        {% if grid.columns.showtitles %}
-            {{ grid_titles(grid) }}
-        {% endif %}
-
-        {% if grid.columns.showfilters %}
-            {{ grid_filters(grid) }}
-        {% endif %}
-
-        {{ grid_rows(grid) }}
-        </table>
-        {{ grid_pager(grid) }}
-    </form>
-</div>
-{% endblock %}
-{# --------------------------------------------------- grid_titles -------------------------------------------------- #}
-{% block grid_titles %}
-    <tr>
-    {% for column in grid.columns %}
-        {% if (column.visible )%}
-        <th class="title"{% if(column.size > 0) %} style="width:{{column.size}}px"{% endif %}>
-        {% if (column.sortable) %}
-            <a href="{{ grid_url('order', grid, column) }}" title="Order by {{ column.title }}">{{ column.title }}</a>
-            {% if column.order == 'asc' %}
-                <div class="sort_up"></div>
-            {% elseif column.order == 'desc' %}
-                <div class="sort_down"></div>
-            {% endif %}
-        {% else %}
-            {{ column.title }}
-        {% endif %}
-        </th>
-        {% endif %}
-    {% endfor %}
-    </tr>
-{% endblock %}
-{# -------------------------------------------------- grid_filters -------------------------------------------------- #}
-{% block grid_filters %}
-    <tr>
-    {% for column in grid.columns %}
-        {% if (column.visible )%}
-        <th>{{ column.filter|raw }}</th>
-        {% endif %}
-    {% endfor %}
-    </tr>
-{% endblock %}
-{# ---------------------------------------------------- grid_rows --------------------------------------------------- #}
-{% block grid_rows %}
-    {% for row in grid.rows %}
-    <tr{% if row.color != '' %} style="background-color:{{row.color}};"{% endif %}>
-    {% for column in grid.columns %}
-        {% if (column.visible )%}
-            {% if loop.last %}
-            <td class="last">{{row.field(column.id)|raw}}</td>
-            {% else %}
-            <td>{{row.field(column.id)|raw}}</td>
-            {% endif %}
-        {% endif %}
-    {% endfor %}
-    </tr>
-    {% endfor %}
-{% endblock %}
-{# ---------------------------------------------------- grid_pager -------------------------------------------------- #}
-{% block grid_pager %}
-<div class="pager">
-Page:
-{% spaceless %}
-<input type="button" class="prev" {% if grid.page <= 0 %}disabled="disabled"{% endif %} value="<" onclick="window.location.href='{{grid_url('page', grid, grid.page-1)}}'"/>
-<input type="text" value="{{grid.page+1}}" size="2" onkeypress="if (event.which == 13)window.location.href='{{grid_url('page', grid)}}'+(parseInt(this.value)-1)"/>
-<input type="button" value=">" class="next" {% if grid.page >= grid.pagecount-1 %}disabled="disabled"{% endif %} onclick="window.location.href='{{grid_url('page', grid, grid.page+1)}}'"/> of {{grid.pagecount}}
-{% endspaceless %}
-, Display <select onchange="window.location.href='{{grid_url('limit', grid)}}'+this.value">
-{% for key, value in grid.limits %}
-    {% if key == grid.currentlimit %}
-        <option value="{{key}}" selected="selected">{{value}}</option>
-    {% else %}
-        <option value="{{key}}">{{value}}</option>
-    {% endif %}
-{% endfor %}
-</select> Items per page
-</div>
-{% endblock %}
-{# ------------------------------------------------ grid_massactions ------------------------------------------------ #}
-{% block grid_massactions %}
-<div class="mass-actions">
-<a href="#">Select Vidible</a> |
-<a href="#">Deselect Vidible</a> |
-<a href="#">Select All</a> | 
-<a href="#">Deselect All</a>
-<div style="float:right">
-Action: <select>
-    <option value="0"></option>
-    <option value="1">Remove</option>
-</select><input type="button" value="Submit Action"/>
-</div>
-</div>
-{% endblock %}
-
-
