--- DataGridBundle/Source/Array.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Sorien\DataGridBundle\Source;
-
-use Sorien\DataGridBundle\Source;
-
-class Array extends Source
-{
-    function prepare()
-    {
-    }
-
-    function execute()
-    {
-    }
-}
