--- DataGridBundle/Source/Csv.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Sorien\DataGridBundle\Source;
-
-use Sorien\DataGridBundle\Source;
-
-class Csv extends Source
-{
-    function prepare()
-    {
-    }
-
-    function execute()
-    {
-    }
-}
