+++ Resources/config/templating_twig.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" ?>
+
+<container xmlns="http://symfony.com/schema/dic/services"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+
+    <parameters>    
+        <parameter key="sorien_datagrid.twig_extention.class">Sorien\DataGridBundle\Twig\DataGridExtention</parameter>
+    </parameters>
+
+    <services>
+        <service id="sorien_datagrid.twig_extention" class="%sorien_datagrid.twig_extention.class%" public="false">
+            <tag name="twig.extension" />
+        </service>
+    </services>
+</container>
