+++ Resources/views/scripts.html.twig
@@ -1,0 +1,24 @@
+{# ----------------------------------------------------- jquery ----------------------------------------------------- #}
+{% block grid_scripts_jquery %}
+{% if ajax %}
+{% endif %}
+
+function selectAll()
+{
+    return $(this).parents('fieldset:eq(0)').find(':checkbox').attr('checked', this.checked);
+}
+
+
+{% endblock %}
+
+{# --------------------------------------------------- prototype ---------------------------------------------------- #}
+{% block grid_scripts_prototype %}
+{% if ajax %}
+{% else %}
+{% endif %}
+{% endblock %}
+
+{# ----------------------------------------------------- native ----------------------------------------------------- #}
+{% block grid_scripts %}
+
+{% endblock %}
