+++ Source/ArraySource.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Sorien\DataGridBundle\Source;
+
+class ArraySource extends Source
+{
+    function prepare()
+    {
+    }
+
+    function execute()
+    {
+    }
+}
