+++ Source/CSVSource.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Sorien\DataGridBundle\Source;
+
+class CSVSource extends Source
+{
+    function prepare()
+    {
+    }
+
+    function execute()
+    {
+    }
+}
