+++ Source/XMLSource.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Sorien\DataGridBundle\Source;
+
+class XMLSource extends Source
+{
+    function prepare()
+    {
+    }
+
+    function execute()
+    {
+    }
+}
